resource/cdk/requirements-dev.txt
aws-cdk-lib==2.180.0
constructs>=10.0.0,<11.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
